+++ dmr/metadata.py
@@ -2,6 +2,7 @@ import dataclasses
 import typing as ty
 from abc import abstractmethod
 from collections.abc import AsyncGenerator, AsyncIterator, Mapping, Set
+from functools import cached_property
 from http import HTTPStatus
 from typing import (  # noqa: WPS235
     TYPE_CHECKING,
@@ -260,10 +261,9 @@ class ResponseSpecProvider:
 
     __slots__ = ()
 
-    @classmethod
     @abstractmethod
     def provide_response_specs(
-        cls,
+        self,
         metadata: 'EndpointMetadata',
         controller_cls: type['Controller[BaseSerializer]'],
         existing_responses: Mapping[HTTPStatus, ResponseSpec],
@@ -276,9 +276,8 @@ class ResponseSpecProvider:
         """
         raise NotImplementedError
 
-    @classmethod
     def _add_new_response(
-        cls,
+        self,
         response: ResponseSpec,
         existing_responses: Mapping[HTTPStatus, ResponseSpec],
     ) -> list[ResponseSpec]:
@@ -385,7 +384,7 @@ class EndpointMetadata:
     parsers: dict[str, 'Parser']
     renderers: dict[str, 'Renderer']
     auth: list['SyncAuth | AsyncAuth'] | None
-    throttling: list['SyncThrottle | AsyncThrottle'] | None
+    throttling: tuple['SyncThrottle | AsyncThrottle', ...] | None
     no_validate_http_spec: frozenset['HttpSpec']
     allowed_http_methods: frozenset[str]
     semantic_responses: bool
@@ -401,6 +400,11 @@ class EndpointMetadata:
     external_docs: 'ExternalDocumentation | None' = None
     callbacks: dict[str, 'Callback | Reference'] | None = None
     servers: list['Server'] | None = None
+
+    def __post_init__(self) -> None:
+        """Trigger cache on start, not on first request."""
+        self.throttling_before_auth
+        self.throttling_after_auth
 
     def collect_response_specs(
         self,
@@ -447,7 +451,7 @@ class EndpointMetadata:
             for response in all_responses
         ]
 
-    def response_spec_providers(self) -> list[type[ResponseSpecProvider]]:
+    def response_spec_providers(self) -> list[ResponseSpecProvider]:
         """
         Determine: from where we should collect response schemas.
 
@@ -467,12 +471,38 @@ class EndpointMetadata:
             return []
 
         return [
-            *[type(spec[0]) for spec in self.component_parsers],
-            *[type(parser) for parser in self.parsers.values()],
-            *[type(renderer) for renderer in self.renderers.values()],
-            *[type(auth) for auth in (self.auth or [])],
-            *[type(throttling) for throttling in (self.throttling or [])],
+            *[spec[0] for spec in self.component_parsers],
+            *[parser for parser in self.parsers.values()],
+            *[renderer for renderer in self.renderers.values()],
+            *[auth for auth in (self.auth or [])],
+            *[throttling for throttling in (self.throttling or [])],
         ]
+
+    @cached_property
+    def throttling_before_auth(
+        self,
+    ) -> tuple['SyncThrottle | AsyncThrottle', ...] | None:
+        return (
+            tuple(
+                throttling
+                for throttling in (self.throttling or ())
+                if throttling.cache_key.runs_before_auth
+            )
+            or None
+        )
+
+    @cached_property
+    def throttling_after_auth(
+        self,
+    ) -> tuple['SyncThrottle | AsyncThrottle', ...] | None:
+        return (
+            tuple(
+                throttling
+                for throttling in (self.throttling or ())
+                if not throttling.cache_key.runs_before_auth
+            )
+            or None
+        )
 
 
 _MetadataT = TypeVar('_MetadataT')