+++ dmr/throttling/algorithms.py
@@ -8,6 +8,9 @@ from dmr.exceptions import TooManyRequestsError
 from dmr.throttling.backends import CachedRateLimit
 
 if TYPE_CHECKING:
+    from dmr.controller import Controller
+    from dmr.endpoint import Endpoint
+    from dmr.serializer import BaseSerializer
     from dmr.throttling import AsyncThrottle, SyncThrottle
 
 
@@ -17,14 +20,14 @@ class BaseThrottleAlgorithm:
     @abc.abstractmethod
     def access(
         self,
+        endpoint: 'Endpoint',
+        controller: 'Controller[BaseSerializer]',
         throttle: 'SyncThrottle | AsyncThrottle',
-        cache_object: 'CachedRateLimit | None',
-        max_requests: int,
-        duration_in_seconds: int,
-    ) -> 'CachedRateLimit': ...
+        cache_object: CachedRateLimit | None,
+    ) -> CachedRateLimit: ...
 
     @abc.abstractmethod
-    def record(self, cache_object: 'CachedRateLimit') -> 'CachedRateLimit': ...
+    def record(self, cache_object: CachedRateLimit) -> CachedRateLimit: ...
 
 
 class SimpleRate(BaseThrottleAlgorithm):
@@ -33,28 +36,30 @@ class SimpleRate(BaseThrottleAlgorithm):
     @override
     def access(
         self,
+        endpoint: 'Endpoint',
+        controller: 'Controller[BaseSerializer]',
         throttle: 'SyncThrottle | AsyncThrottle',
-        cache_object: 'CachedRateLimit | None',
-        max_requests: int,
-        duration_in_seconds: int,
-    ) -> 'CachedRateLimit':
+        cache_object: CachedRateLimit | None,
+    ) -> CachedRateLimit:
         now = int(time.time())
         if cache_object is None or cache_object['reset'] <= now:
             # For this algorithm we use a single history
             # item which is the number of calls:
-            return {'history': [0], 'reset': now + duration_in_seconds}
+            return {'history': [0], 'reset': now + throttle.duration_in_seconds}
 
         requests = cache_object['history'][0]
-        if requests >= max_requests:
+        if requests >= throttle.max_requests:
             raise TooManyRequestsError(
-                headers=throttle.headers(
-                    remaining=max_requests - requests,
+                headers=throttle.collect_response_headers(
+                    endpoint,
+                    controller,
+                    remaining=throttle.max_requests - requests,
                     reset=cache_object['reset'] - now,
                 ),
             )
         return cache_object
 
     @override
-    def record(self, cache_object: 'CachedRateLimit') -> 'CachedRateLimit':
+    def record(self, cache_object: CachedRateLimit) -> CachedRateLimit:
         cache_object['history'][0] += 1
         return cache_object