+++ dmr/throttling/backends.py
@@ -8,6 +8,7 @@ from dmr.settings import default_parser, default_renderer
 
 if TYPE_CHECKING:
     from dmr.controller import Controller
+    from dmr.endpoint import Endpoint
     from dmr.serializer import BaseSerializer
 
 
@@ -24,25 +25,28 @@ class BaseThrottleBackend:
     @abc.abstractmethod
     def get(
         self,
-        cache_key: str,
+        endpoint: 'Endpoint',
         controller: 'Controller[BaseSerializer]',
+        cache_key: str,
     ) -> CachedRateLimit | None:
         raise NotImplementedError
 
     @abc.abstractmethod
     async def aget(
         self,
-        cache_key: str,
+        endpoint: 'Endpoint',
         controller: 'Controller[BaseSerializer]',
+        cache_key: str,
     ) -> CachedRateLimit | None:
         raise NotImplementedError
 
     @abc.abstractmethod
     def set(
         self,
+        endpoint: 'Endpoint',
+        controller: 'Controller[BaseSerializer]',
         cache_key: str,
         cache_object: CachedRateLimit,
-        controller: 'Controller[BaseSerializer]',
         *,
         ttl_seconds: int,
     ) -> None:
@@ -51,16 +55,17 @@ class BaseThrottleBackend:
     @abc.abstractmethod
     async def aset(
         self,
+        endpoint: 'Endpoint',
+        controller: 'Controller[BaseSerializer]',
         cache_key: str,
         cache_object: CachedRateLimit,
-        controller: 'Controller[BaseSerializer]',
         *,
         ttl_seconds: int,
     ) -> None:
         raise NotImplementedError
 
 
-class DjangoCacheBaseThrottleBackend(BaseThrottleBackend):
+class DjangoCache(BaseThrottleBackend):
     __slots__ = ('_cache',)
 
     def __init__(self, cache_name: str = DEFAULT_CACHE_ALIAS) -> None:
@@ -69,55 +74,59 @@ class DjangoCacheBaseThrottleBackend(BaseThrottleBackend):
     @override
     def get(
         self,
-        cache_key: str,
+        endpoint: 'Endpoint',
         controller: 'Controller[BaseSerializer]',
+        cache_key: str,
     ) -> CachedRateLimit | None:
         stored_cache = self._cache.get(cache_key)
-        return self._load_cache(stored_cache, controller)
+        return self._load_cache(controller, stored_cache)
 
     @override
     async def aget(
         self,
-        cache_key: str,
+        endpoint: 'Endpoint',
         controller: 'Controller[BaseSerializer]',
+        cache_key: str,
     ) -> CachedRateLimit | None:
         stored_cache = await self._cache.aget(cache_key)
-        return self._load_cache(stored_cache, controller)
+        return self._load_cache(controller, stored_cache)
 
     @override
     def set(
         self,
+        endpoint: 'Endpoint',
+        controller: 'Controller[BaseSerializer]',
         cache_key: str,
         cache_object: CachedRateLimit,
-        controller: 'Controller[BaseSerializer]',
         *,
         ttl_seconds: int,
     ) -> None:
         self._cache.set(
             cache_key,
-            self._dump_cache(cache_object, controller),
+            self._dump_cache(controller, cache_object),
             timeout=ttl_seconds,
         )
 
     @override
     async def aset(
         self,
+        endpoint: 'Endpoint',
+        controller: 'Controller[BaseSerializer]',
         cache_key: str,
         cache_object: CachedRateLimit,
-        controller: 'Controller[BaseSerializer]',
         *,
         ttl_seconds: int,
     ) -> None:
         await self._cache.aset(
             cache_key,
-            self._dump_cache(cache_object, controller),
+            self._dump_cache(controller, cache_object),
             timeout=ttl_seconds,
         )
 
     def _load_cache(
         self,
-        stored_cache: bytes | None,
         controller: 'Controller[BaseSerializer]',
+        stored_cache: bytes | None,
     ) -> CachedRateLimit | None:
         if stored_cache is None:
             return None
@@ -131,8 +140,8 @@ class DjangoCacheBaseThrottleBackend(BaseThrottleBackend):
 
     def _dump_cache(
         self,
-        cache_object: CachedRateLimit,
         controller: 'Controller[BaseSerializer]',
+        cache_object: CachedRateLimit,
     ) -> bytes:
         return controller.serializer.serialize(
             cache_object,