+++ dmr/throttling/base.py
@@ -1,10 +1,11 @@
 import asyncio
 import enum
 import threading
-from collections.abc import Callable, Mapping
+from collections.abc import Iterable, Mapping
 from http import HTTPStatus
-from typing import TYPE_CHECKING, ClassVar, TypeAlias, final
+from typing import TYPE_CHECKING, Literal, final, overload
 
+from django.http import HttpRequest
 from typing_extensions import override
 
 from dmr.exceptions import TooManyRequestsError
@@ -16,9 +17,14 @@ from dmr.throttling.algorithms import (
 )
 from dmr.throttling.backends import (
     BaseThrottleBackend,
-    DjangoCacheBaseThrottleBackend,
+    DjangoCache,
 )
-from dmr.throttling.cache_keys import remote_address
+from dmr.throttling.cache_keys import BaseThrottleCacheKey, RemoteAddr
+from dmr.throttling.headers import (
+    BaseResponseHeadersProvider,
+    RetryAfter,
+    XRateLimit,
+)
 
 if TYPE_CHECKING:
     from dmr.controller import Controller
@@ -35,41 +41,37 @@ class Rate(enum.IntEnum):
     day = 24 * hour
 
 
-_CacheKey: TypeAlias = Callable[
-    ['Endpoint', 'Controller[BaseSerializer]'],
-    str | None,
-]
-
-
 class _BaseThrottle(ResponseSpecProvider):
-    header_prefix: ClassVar[str] = ''
-
     __slots__ = (
         '_algorithm',
         '_async_lock',
         '_backend',
-        '_cache_key',
-        '_duration_in_seconds',
-        '_max_requests',
+        '_response_headers',
         '_sync_lock',
+        'cache_key',
+        'duration_in_seconds',
+        'max_requests',
     )
 
     def __init__(
         self,
         max_requests: int,
-        rate: Rate | int,
+        durantion_in_seconds: Rate | int,
         *,
-        cache_key: _CacheKey | None = None,
-        number_of_proxies: int | None = None,
+        cache_key: BaseThrottleCacheKey | None = None,
         backend: BaseThrottleBackend | None = None,
         algorithm: BaseThrottleAlgorithm | None = None,
+        response_headers: Iterable[BaseResponseHeadersProvider] | None = None,
     ) -> None:
-        self._max_requests = max_requests
-        self._duration_in_seconds = rate
+        self.max_requests = max_requests
+        self.duration_in_seconds = int(durantion_in_seconds)
         # Default implementations of the logical parts:
-        self._cache_key = cache_key or remote_address
-        self._backend = backend or DjangoCacheBaseThrottleBackend()
+        self.cache_key = cache_key or RemoteAddr()
+        self._backend = backend or DjangoCache()
         self._algorithm = algorithm or SimpleRate()
+        self._response_headers = (
+            [XRateLimit(), RetryAfter()] if response_headers is None else []
+        )
         # Locks:
         self._sync_lock = threading.Lock()
         self._async_lock = asyncio.Lock()
@@ -79,70 +81,64 @@ class _BaseThrottle(ResponseSpecProvider):
         endpoint: 'Endpoint',
         controller: 'Controller[BaseSerializer]',
     ) -> str | None:
-        cache_key = self._cache_key(endpoint, controller)
+        cache_key = self.cache_key(endpoint, controller)
         if cache_key is None:
             return None
 
         metadata = endpoint.metadata
         backend_name = type(self._backend).__qualname__
         algorith_name = type(self._algorithm).__qualname__
-        cache_key_name = getattr(self._cache_key, '__qualname__', '')
+        cache_key_name = type(self.cache_key).__qualname__
         return (
-            f'{metadata.operation_id}_{metadata.method}_'
-            f'{backend_name}_{algorith_name}_'
-            f'{self._max_requests}_{self._duration_in_seconds}_'
-            f'{cache_key_name}_{cache_key}'
+            f'{metadata.operation_id}::{metadata.method}::'
+            f'{backend_name}::{algorith_name}::'
+            f'{cache_key_name}::{cache_key}::'
+            f'{self.max_requests}::{self.duration_in_seconds}'
         )
 
-    def headers(self, remaining: int, reset: int) -> dict[str, str]:
-        prefix = self.header_prefix
-        return {
-            f'{prefix}RateLimit-Limit': str(self._max_requests),
-            f'{prefix}RateLimit-Remaining': str(remaining),
-            f'{prefix}RateLimit-Reset': str(reset),
-        }
-
     @override
-    @classmethod
     def provide_response_specs(
-        cls,
+        self,
         metadata: EndpointMetadata,
         controller_cls: type['Controller[BaseSerializer]'],
         existing_responses: Mapping[HTTPStatus, ResponseSpec],
     ) -> list[ResponseSpec]:
         """Provides responses that can happen when throttle triggers."""
-        return cls._add_new_response(
+        return self._add_new_response(
             ResponseSpec(
                 controller_cls.error_model,
                 status_code=TooManyRequestsError.status_code,
-                headers=cls.headers_spec(),
+                headers=self._headers_spec(),
                 description='Raised when throttling rate was hit',
             ),
             existing_responses,
         )
 
-    @classmethod
-    def headers_spec(cls) -> dict[str, HeaderSpec]:
-        return {
-            f'{cls.header_prefix}RateLimit-Limit': HeaderSpec(
-                description=(
-                    'The maximum number of requests permitted '
-                    'in the current time window'
+    def collect_response_headers(
+        self,
+        endpoint: 'Endpoint',
+        controller: 'Controller[BaseSerializer]',
+        remaining: int,
+        reset: int,
+    ) -> dict[str, str]:
+        response_headers = {}
+        for header_provider in self._response_headers:
+            response_headers.update(
+                header_provider.response_headers(
+                    endpoint,
+                    controller,
+                    self,
+                    remaining,
+                    reset,
                 ),
-            ),
-            f'{cls.header_prefix}RateLimit-Remaining': HeaderSpec(
-                description=(
-                    'The number of requests remaining '
-                    'in the current time window'
-                ),
-            ),
-            f'{cls.header_prefix}RateLimit-Reset': HeaderSpec(
-                description=(
-                    'The number of seconds until the current '
-                    'rate limit window resets'
-                ),
-            ),
-        }
+            )
+        return response_headers
+
+    def _headers_spec(self) -> dict[str, HeaderSpec]:
+        headers_spec = {}
+        for header_provider in self._response_headers:
+            headers_spec.update(header_provider.provide_headers_specs())
+        return headers_spec
 
 
 class SyncThrottle(_BaseThrottle):
@@ -163,31 +159,30 @@ class SyncThrottle(_BaseThrottle):
         if you want to change the return code, for example,
         when some data is missing or has wrong format.
         """
-        cache_key = self.get_cache_key(
-            endpoint,
-            controller,
-        )
+        cache_key = self.get_cache_key(endpoint, controller)
         if cache_key is None:
             return
-        self.check(cache_key, controller)
+        self.check(endpoint, controller, cache_key)
 
     def check(
         self,
-        cache_key: str,
+        endpoint: 'Endpoint',
         controller: 'Controller[BaseSerializer]',
+        cache_key: str,
     ) -> None:
         with self._sync_lock:
             cache_object = self._algorithm.access(
+                endpoint,
+                controller,
                 self,
-                self._backend.get(cache_key, controller),
-                self._max_requests,
-                int(self._duration_in_seconds),
+                self._backend.get(endpoint, controller, cache_key),
             )
             self._backend.set(
+                endpoint,
+                controller,
                 cache_key,
                 self._algorithm.record(cache_object),
-                controller,
-                ttl_seconds=self._duration_in_seconds,
+                ttl_seconds=self.duration_in_seconds,
             )
 
 
@@ -209,29 +204,61 @@ class AsyncThrottle(_BaseThrottle):
         if you want to change the return code, for example,
         when some data is missing or has wrong format.
         """
-        cache_key = self.get_cache_key(
-            endpoint,
-            controller,
-        )
+        cache_key = self.get_cache_key(endpoint, controller)
         if cache_key is None:
             return
-        await self.check(cache_key, controller)
+        await self.check(endpoint, controller, cache_key)
 
     async def check(
         self,
-        cache_key: str,
+        endpoint: 'Endpoint',
         controller: 'Controller[BaseSerializer]',
+        cache_key: str,
     ) -> None:
         async with self._async_lock:
             cache_object = self._algorithm.access(
+                endpoint,
+                controller,
                 self,
-                await self._backend.aget(cache_key, controller),
-                self._max_requests,
-                int(self._duration_in_seconds),
+                await self._backend.aget(endpoint, controller, cache_key),
             )
             await self._backend.aset(
+                endpoint,
+                controller,
                 cache_key,
                 self._algorithm.record(cache_object),
-                controller,
-                ttl_seconds=self._duration_in_seconds,
+                ttl_seconds=self.duration_in_seconds,
             )
+
+
+@overload
+def request_throttling(
+    request: HttpRequest,
+    *,
+    strict: Literal[True],
+) -> tuple[SyncThrottle | AsyncThrottle, ...]: ...
+
+
+@overload
+def request_throttling(
+    request: HttpRequest,
+    *,
+    strict: bool = False,
+) -> tuple[SyncThrottle | AsyncThrottle, ...] | None: ...
+
+
+def request_throttling(
+    request: HttpRequest,
+    *,
+    strict: bool = False,
+) -> tuple[SyncThrottle | AsyncThrottle, ...] | None:
+    """
+    Return the tuple of throttling instances that were for this request.
+
+    When *strict* is passed and *request* has no throttling,
+    we raise :exc:`AttributeError`.
+    """
+    throttling = getattr(request, '__dmr_throttling__', None)
+    if throttling is None and strict:
+        raise AttributeError('__dmr_throttling__')
+    return throttling