+++ dmr/throttling/cache_keys.py
@@ -1,4 +1,8 @@
+import abc
+import dataclasses
 from typing import TYPE_CHECKING
+
+from typing_extensions import override
 
 if TYPE_CHECKING:
     from dmr.controller import Controller
@@ -6,8 +10,27 @@ if TYPE_CHECKING:
     from dmr.serializer import BaseSerializer
 
 
-def remote_address(
-    endpoint: 'Endpoint',
-    controller: 'Controller[BaseSerializer]',
-) -> str | None:
-    return controller.request.META.get('REMOTE_ADDR')
+class BaseThrottleCacheKey:
+    runs_before_auth: bool
+    name: str
+
+    @abc.abstractmethod
+    def __call__(
+        self,
+        endpoint: 'Endpoint',
+        controller: 'Controller[BaseSerializer]',
+    ) -> str | None: ...
+
+
+@dataclasses.dataclass(slots=True, frozen=True)
+class RemoteAddr(BaseThrottleCacheKey):
+    runs_before_auth: bool = True
+    name: str = 'RemoteAddr'
+
+    @override
+    def __call__(
+        self,
+        endpoint: 'Endpoint',
+        controller: 'Controller[BaseSerializer]',
+    ) -> str | None:
+        return controller.request.META.get('RemoteAddr')