+++ tests/test_unit/test_throttling/test_throttling_custom_header.py
@@ -10,6 +10,7 @@ from dmr import Controller, modify
 from dmr.plugins.pydantic import PydanticSerializer
 from dmr.test import DMRAsyncRequestFactory, DMRRequestFactory
 from dmr.throttling import AsyncThrottle, Rate, SyncThrottle
+from dmr.throttling.headers import RateLimitIETFDraft, RetryAfter, XRateLimit
 
 
 class _XSyncThrottle(SyncThrottle):
@@ -22,7 +23,7 @@ class _SyncEndpointController(Controller[PydanticSerializer]):
         return 'inside'
 
 
-def test_throttle_sync_per_endpoint(
+def test_throttle_sync_x_prefix(
     dmr_rf: DMRRequestFactory,
     freezer: FrozenDateTimeFactory,
 ) -> None:
@@ -46,6 +47,7 @@ def test_throttle_sync_per_endpoint(
         'X-RateLimit-Limit': '1',
         'X-RateLimit-Remaining': '0',
         'X-RateLimit-Reset': '1',
+        'Retry-After': '1',
         'Content-Type': 'application/json',
     }
     assert json.loads(response.content) == snapshot({
@@ -67,7 +69,7 @@ class _AsyncController(
 
 
 @pytest.mark.asyncio
-async def test_throttle_async_per_endpoint(
+async def test_throttle_async_x_prefix(
     dmr_async_rf: DMRAsyncRequestFactory,
 ) -> None:
     """Ensures custom throttle prefix async."""
@@ -90,8 +92,59 @@ async def test_throttle_async_per_endpoint(
         'X-RateLimit-Limit': '1',
         'X-RateLimit-Remaining': '0',
         'X-RateLimit-Reset': '1',
+        'Retry-After': '1',
         'Content-Type': 'application/json',
     }
     assert json.loads(response.content) == snapshot({
         'detail': [{'msg': 'Too many requests', 'type': 'ratelimit'}],
     })
+
+
+class _SyncNoHeadersController(Controller[PydanticSerializer]):
+    @modify(throttling=[SyncThrottle(1, Rate.second, response_headers=())])
+    def get(self) -> str:
+        return 'inside'
+
+
+def test_throttle_sync_no_headers(
+    dmr_rf: DMRRequestFactory,
+    freezer: FrozenDateTimeFactory,
+) -> None:
+    """Ensures custom headers rules."""
+    # First will pass:
+    request = dmr_rf.get('/whatever/')
+    response = _SyncNoHeadersController.as_view()(request)
+    assert isinstance(response, HttpResponse)
+    assert response.status_code == HTTPStatus.OK, response.content
+    assert response.headers == {'Content-Type': 'application/json'}
+    assert json.loads(response.content) == 'inside'
+
+    # This will fail:
+    request = dmr_rf.get('/whatever/')
+    response = _SyncNoHeadersController.as_view()(request)
+    assert isinstance(response, HttpResponse)
+    assert response.status_code == HTTPStatus.TOO_MANY_REQUESTS, (
+        response.content
+    )
+    assert response.headers == {'Content-Type': 'application/json'}
+    assert json.loads(response.content) == snapshot({
+        'detail': [{'msg': 'Too many requests', 'type': 'ratelimit'}],
+    })
+
+
+class _SyncAllHeadersController(Controller[PydanticSerializer]):
+    @modify(
+        throttling=[
+            SyncThrottle(
+                1,
+                Rate.second,
+                response_headers=[
+                    RetryAfter(),
+                    XRateLimit(),
+                    RateLimitIETFDraft(),
+                ],
+            ),
+        ],
+    )
+    def get(self) -> str:
+        return 'inside'